pyproject.toml
[project]
authors = [{ name = "Konstantin Borovik", email = "kb@lab5.ca" }]
dependencies = [
  "argcomplete>=3.5.1",
  "google-cloud-iam>=2.16.0",
  "google-cloud-service-usage>=1.11.1",
  "rich>=14.0.0",
]
description = "Query GCP IAM roles and permissions"
name = "gcp-iam-roles"
readme = "README.md"
requires-python = "~=3.13.3"
version = "2025.06.13.post2322"

[project.scripts]
gcp-iam-roles = "gcp_iam_roles:cli"

[build-system]
build-backend = "hatchling.build"
requires = ["hatchling"]

[dependency-groups]
dev = ["ruff>=0.7.3"]

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = ["C90", "F", "N", "N", "PL", "RUF", "SIM", "UP"]

[tool.ruff.lint.pydocstyle]
convention = "google"
